feat: create lesson
