+++ lms-service/src/main/java/com/lmsservice/entity/Lesson.java
@@ -11,6 +11,7 @@ import lombok.experimental.FieldDefaults;
 @Entity
 @Getter
 @Setter
+@Builder
 @NoArgsConstructor
 @AllArgsConstructor
 @FieldDefaults(level = AccessLevel.PRIVATE)
